requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.4
discord==2.3.2
discord.py==2.4.0
frozenlist==1.4.1
gTTS==2.5.1
idna==3.7
multidict==6.0.5
pycparser==2.22
pydub==0.25.1
PyNaCl==1.5.0
pypiwin32==223
python-dotenv==1.0.1
pyttsx3==2.90
pywin32==306
requests==2.32.3
urllib3==2.2.2
yarl==1.9.4